+++ packages/__tests__/src/3-runtime-html/binding-behavior.spec.ts
@@ -1,6 +1,8 @@
 import { resolve } from '@aurelia/kernel';
 import {
   Scope,
+} from '@aurelia/runtime';
+import {
   type BindingBehaviorInstance,
   type IBinding,
   bindingBehavior,
