+++ packages/__tests__/src/3-runtime-html/promise.spec.ts
@@ -19,6 +19,8 @@ import {
 } from '@aurelia/kernel';
 import {
   Scope,
+} from '@aurelia/runtime';
+import {
   type BindingBehaviorInstance,
   type IBinding,
   valueConverter,
