+++ packages/__tests__/src/3-runtime-html/switch.spec.ts
@@ -13,6 +13,8 @@ import {
 } from '@aurelia/kernel';
 import {
   Scope,
+} from '@aurelia/runtime';
+import {
   AuSlot,
   type BindingBehaviorInstance,
   type IBinding,
