+++ packages/aurelia/src/api-test.ts
@@ -1,0 +1,8 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+
+import Aurelia from '.';
+
+function testAureliaApp() {
+  void Aurelia.app({ host: document.body, component: class { } }).start();
+  void Aurelia.enhance({ host: document.body, component: class { } });
+}
