+++ packages/i18n/src/df/date-format-binding-behavior.ts
@@ -1,6 +1,6 @@
+import { type Scope } from '@aurelia/runtime';
+import { type BindingBehaviorStaticAuDefinition, type BindingBehaviorInstance, } from '@aurelia/runtime-html';
 import { type BindingWithBehavior, createIntlFormatValueConverterExpression, ValueConverters, behaviorTypeName } from '../utils';
-
-import { type Scope, type BindingBehaviorStaticAuDefinition, type BindingBehaviorInstance, } from '@aurelia/runtime-html';
 
 export class DateFormatBindingBehavior implements BindingBehaviorInstance {
   public static readonly $au: BindingBehaviorStaticAuDefinition = {
