+++ packages/i18n/src/rt/relative-time-binding-behavior.ts
@@ -1,4 +1,5 @@
-import { type BindingBehaviorInstance, type BindingBehaviorStaticAuDefinition, type Scope } from '@aurelia/runtime-html';
+import { type Scope } from '@aurelia/runtime';
+import { type BindingBehaviorInstance, type BindingBehaviorStaticAuDefinition } from '@aurelia/runtime-html';
 import { ValueConverters, behaviorTypeName, createIntlFormatValueConverterExpression, type BindingWithBehavior } from '../utils';
 
 export class RelativeTimeBindingBehavior implements BindingBehaviorInstance {
