+++ packages/runtime-html/src/resources/binding-behaviors/signals.ts
@@ -1,9 +1,8 @@
+import { resolve } from '@aurelia/kernel';
+import { type Scope, type ISubscriber } from '@aurelia/runtime';
 import { ISignaler } from '../../signaler';
 import { type BindingBehaviorInstance, type BindingBehaviorStaticAuDefinition, behaviorTypeName } from '../binding-behavior';
 import { addSignalListener, removeSignalListener } from '../../utilities';
-import { type ISubscriber } from '@aurelia/runtime';
-import { resolve } from '@aurelia/kernel';
-import { type Scope } from '../../binding/scope';
 import { ErrorNames, createMappedError } from '../../errors';
 import { IBinding } from '../../binding/interfaces-bindings';
