+++ packages/runtime-html/src/resources/binding-behaviors/throttle.ts
@@ -1,6 +1,6 @@
-import { isString, IPlatform, type IDisposable, emptyArray, resolve } from '@aurelia/kernel';
 import { TaskQueue } from '@aurelia/platform';
-import { type Scope } from '../../binding/scope';
+import { isString, IPlatform, type IDisposable, emptyArray, resolve } from '@aurelia/kernel';
+import { type Scope } from '@aurelia/runtime';
 import { type BindingBehaviorInstance, BindingBehaviorStaticAuDefinition, behaviorTypeName } from '../binding-behavior';
 import { type IBinding, type IRateLimitOptions } from '../../binding/interfaces-bindings';
