+++ packages/runtime-html/src/resources/template-controllers/with.ts
@@ -1,9 +1,9 @@
-import { Scope } from '../../binding/scope';
+import { resolve } from '@aurelia/kernel';
+import { Scope } from '@aurelia/runtime';
 import { IRenderLocation } from '../../dom';
 import { IViewFactory } from '../../templating/view';
 import { CustomAttributeStaticAuDefinition, attrTypeName } from '../custom-attribute';
 import type { ICustomAttributeController, ICustomAttributeViewModel, IHydratedController, IHydratedParentController, ControllerVisitor } from '../../templating/controller';
-import { resolve } from '@aurelia/kernel';
 
 export class With implements ICustomAttributeViewModel {
   public static readonly $au: CustomAttributeStaticAuDefinition = {
